pushing

+++ 2017/assignment2/color_utils.py
@@ -290,7 +290,7 @@ def decode(data_l, conv8_313, rebalance=1):
   data_ab = np.dot(class8_313_rh, cc)
   data_ab = resize(data_ab, (height, width))
   img_lab = np.concatenate((data_l, data_ab), axis=-1)
-  img_rgb = color.lab2rgb(img_lab)
+  img_rgb = color.lab2rgb(img_lab).astype( dtype = np.float32 )
 
   return img_rgb
